Fix scripting against SQL running in a docker container.
Running against docker, you cannot user the server name from the connection string when composing a SMO Urn instance. The result is the SMO Urn won't resolve since the docker server name must be used, however you cannot use the docker server name in the connection string. The fix is to connect to the server with the passed connection string and resolve the server name used for SMO Urn.