Differentiate External and FileTable types in object explorer list - Add (External) and (File Table) to the labels - Pre-fetch these properties to avoid perf hit
Note on testing: These are both quite hard to initialize, so adding tests to the integration tests would require always having a server with Polybase enabled / FileTable enabled. Given this I feel it should be OK to use manual testing. Long term, we should have unit-level tests that cover the functionality or figure out how to have a perfect pre-configured server and a set of tests to run the breadth of this logic against it.